lambda_backend/requirements.txt
PyJWT==2.8.0
boto3==1.34.131
